requirements.txt
pandas
python-dotenv
openai
requests
fastapi
uvicorn
